Combine Enums.hpp and Structs.hpp into Types.hpp
Also, rename Structs.ipp to Types.ipp
