+++ Sts1CobcSw/Edu/Edu.hpp
@@ -1,8 +1,7 @@
 #pragma once
 
 
-#include <Sts1CobcSw/Edu/Enums.hpp>
-#include <Sts1CobcSw/Edu/Structs.hpp>
+#include <Sts1CobcSw/Edu/Types.hpp>
 #include <Sts1CobcSw/Outcome/Outcome.hpp>
 #include <Sts1CobcSw/Serial/Byte.hpp>
